move get_create_display_fun to nilearn.plotting.displays._utils
