+++ nilearn/plotting/displays/_slicers.py
@@ -17,6 +17,7 @@ from nilearn.image import get_data, new_img_like, reorder_img
 from nilearn.image.resampling import get_bounds, get_mask_bounds
 from nilearn.plotting._utils import check_threshold_not_negative
 from nilearn.plotting.displays import CutAxes
+from nilearn.plotting.displays._utils import get_create_display_fun
 from nilearn.plotting.displays._axes import coords_3d_to_2d
 from nilearn.plotting.edge_detect import edge_map
 from nilearn.plotting.find_cuts import find_cut_slices, find_xyz_cut_coords
@@ -2211,18 +2212,3 @@ def get_slicer(display_mode):
 
     """
     return get_create_display_fun(display_mode, SLICERS)
-
-
-def get_create_display_fun(display_mode, class_dict):
-    """Help for functions \
-    :func:`~nilearn.plotting.displays.get_slicer` and \
-    :func:`~nilearn.plotting.displays.get_projector`.
-    """
-    try:
-        return class_dict[display_mode].init_with_figure
-    except KeyError:
-        message = (
-            f"{display_mode} is not a valid display_mode. "
-            f"Valid options are {sorted(class_dict.keys())}"
-        )
-        raise ValueError(message)
